cloud_functions/purge_all_for_paper/src/requirements.txt
functions-framework==3.*
arxiv-base @ git+https://github.com/arXiv/arxiv-base.git@fa04631#egg=arxiv_base
google-cloud-logging